backend/realtime_vlm/xai_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiohttp==3.9.3
python-multipart==0.0.7
Pillow==10.3.0
numpy==1.26.4
